requirements.txt
# Core Data Science Libraries
pandas==2.1.4
numpy==1.26.3
matplotlib==3.8.2
seaborn==0.13.1

# Jupyter
jupyter==1.0.0
ipykernel==6.29.0
notebook==7.0.6

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Data Processing
openpyxl==3.1.2
xlrd==2.0.1

# Geospatial (optional, if you need mapping)
geopandas==0.14.2
folium==0.15.1

# API & Web
requests==2.31.0
python-dotenv==1.0.0

# Machine Learning (will add more in Phase 4)
scikit-learn==1.4.0
xgboost==2.0.3

# Utilities
tqdm==4.66.1
python-dateutil==2.8.2